+++ peer/pendingheap/list.go
@@ -25,7 +25,8 @@ import (
 	"time"
 
 	"go.uber.org/yarpc/api/peer"
-	"go.uber.org/yarpc/peer/peerlist/v2"
+	"go.uber.org/yarpc/peer/abstractlist"
+	"go.uber.org/zap"
 )
 
 type listConfig struct {
@@ -34,6 +35,7 @@ type listConfig struct {
 	failFast bool
 	seed     int64
 	nextRand func(int) int
+	logger   *zap.Logger
 }
 
 var defaultListConfig = listConfig{
@@ -64,6 +66,13 @@ func Seed(seed int64) ListOption {
 	}
 }
 
+// Logger provides an optional logger for the list.
+func Logger(logger *zap.Logger) ListOption {
+	return func(c *listConfig) {
+		c.logger = logger
+	}
+}
+
 // FailFast indicates that the peer list should not wait for a peer to become
 // available when choosing a peer.
 //
@@ -82,14 +91,15 @@ func New(transport peer.Transport, opts ...ListOption) *List {
 		o(&cfg)
 	}
 
-	plOpts := []peerlist.ListOption{
-		peerlist.Capacity(cfg.capacity),
+	plOpts := []abstractlist.Option{
+		abstractlist.Capacity(cfg.capacity),
+		abstractlist.Logger(cfg.logger),
 	}
 	if !cfg.shuffle {
-		plOpts = append(plOpts, peerlist.NoShuffle())
+		plOpts = append(plOpts, abstractlist.NoShuffle())
 	}
 	if cfg.failFast {
-		plOpts = append(plOpts, peerlist.FailFast())
+		plOpts = append(plOpts, abstractlist.FailFast())
 	}
 
 	nextRandFn := nextRand(cfg.seed)
@@ -99,7 +109,7 @@ func New(transport peer.Transport, opts ...ListOption) *List {
 	}
 
 	return &List{
-		List: peerlist.New(
+		List: abstractlist.New(
 			"fewest-pending-requests",
 			transport,
 			&pendingHeap{
@@ -110,9 +120,9 @@ func New(transport peer.Transport, opts ...ListOption) *List {
 	}
 }
 
-// List is a PeerList which rotates which peers are to be selected in a circle
+// List is a peer list which rotates which peers are to be selected in a circle
 type List struct {
-	*peerlist.List
+	*abstractlist.List
 }
 
 // nextRand is a convenience function for creating a new rand.Rand from a given