+++ transport/http/close_idle_go112.go
@@ -18,31 +18,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package peer
+// +build go1.12
 
-// Subscriber listens to changes of a Peer over time.
-type Subscriber interface {
-	// The Peer Notifies the Subscriber when its status changes (e.g.
-	// connections status, pending requests)
-	//
-	// NotifyStatusChanged may call methods of the corresponding Peer,
-	// particularly Status.
-	// So, subscriber methods must not be called while holding a lock on the
-	// Transport.
-	NotifyStatusChanged(Identifier)
-}
+package http
 
-// Transport manages Peers across different Subscribers.
-//
-// A Subscriber will request a Peer for a specific PeerIdentifier and the
-// Transport has the ability to create a new Peer or return an existing one.
-//
-// RetainPeer and ReleasePeer must not call or synchronize with goroutines that
-// call methods of the Subscriber.
-type Transport interface {
-	// Get or create a Peer for the Subscriber
-	RetainPeer(Identifier, Subscriber) (Peer, error)
+import "net/http"
 
-	// Unallocate a peer from the Subscriber
-	ReleasePeer(Identifier, Subscriber) error
+// Once YARPC only supports Go1.12+, we can inline this into the callers.
+func closeIdleConnections(client *http.Client) {
+	client.CloseIdleConnections()
 }